feature: loading character also sets visible items
